Keep old versions of krbtgt keys in a keytab
Normally, when a krbtgt key is changed, all tickets for that service are
invalidated because the kdc database only keeps one key at a time.

The tgtextract program copies all krbtgt keys into a keytab, leaving them
encrypted with the master key if one is present.

The main kdc process will now look keys up in this keytab if an incoming
request uses a kvno other than the one the database contains.